feat: match i/o to the spec

+++ src/main/java/io/github/GrbavaCigla/core/Constants.java
@@ -6,8 +6,8 @@ public class Constants {
     public static final int WINDOW_HEIGHT = 700;
     public static final int WINDOW_WIDTH = 1000;
     public static final int AIRPORT_MARKER_SIZE = 14;
-    public static final float AIRPORT_X_LIMIT = 180.0f;
-    public static final float AIRPORT_Y_LIMIT = 90.0f;
+    public static final int AIRPORT_X_LIMIT = 180;
+    public static final int AIRPORT_Y_LIMIT = 90;
     public static final int SIM_STEP_MINUTES = 10;
     public static final int SIM_MICROSTEP_MINUTES = 2;
     public static final int SIM_REAL_MICROSTEP_MS = 200;
